requirements.txt
streamlit>=1.31.0
pandas
scikit-learn
joblib
google-cloud-texttospeech
anthropic
